+++ ecs/query_bench_test.go
@@ -48,9 +48,9 @@ func BenchmarkPosVelQueryUnsafe_1000(b *testing.B) {
 	mapper := NewMap2[Position, Velocity](&world)
 	mapper.NewBatch(n, &Position{}, &Velocity{X: 1, Y: 0})
 
-	filter := NewFilter(posID, velID)
+	filter := NewFilter(&world, posID, velID)
 	for b.Loop() {
-		query := filter.Query(&world)
+		query := filter.Query()
 		for query.Next() {
 			pos := (*Position)(query.Get(posID))
 			vel := (*Velocity)(query.Get(velID))